reorganize code into proper modules
