--- src/main.py
@@ -1,4 +1,0 @@
-from tools_aibg import get_report
-
-if __name__ == "__main__":
-    get_report()
