+++ src/tools_aibg/explorer.py
@@ -1,6 +1,7 @@
 import re
 import pandas as pd
-from typing import TextIO
+import sys
+from typing import Union, TextIO, List
 
 
 class DiaguardImportReader():
@@ -85,3 +86,88 @@ class DiaguardImportReader():
                 i = self.process_entry(content, i+1)
             else:
                 i = i+1
+
+
+
+class Explorer():
+    """Explorer class for obtaining and filtering the entry dataframe"""
+    def __init__(self, f: TextIO):
+        """Constructs the entry dataframe, sorted by ascending datetime"""
+        self.original_df = DiaguardImportReader(f).df
+        self.original_df.sort_values(by="date", ascending=True, inplace=True)
+        self.df = self.original_df.copy()
+
+    def count(self):
+        """Count number of entries"""
+        return self.df.count().max()
+
+    def groupby_hour(self):
+        """Returns df grouped by hour of the day"""
+        return self.df.groupby(self.df["date"].dt.hour)
+
+    def groupby_day(self):
+        """Returns df grouped by date without hour"""
+        return self.df.groupby(self.df["date"].dt.date)
+
+    def groupby_weekday(self):
+        """Returns df grouped by day of the week"""
+        return self.df.groupby(self.df["date"].dt.day_name())
+
+    # filters select data from df and return the explorer itself
+    # (so you can do explorer.glucose(min=70, max=100).carbs(min=5, max=70).df)
+    def reset_df(self):
+        """go back to original df"""
+        self.df = self.original_df.copy()
+        return self
+
+    def col_lims(self, column, lower_bound=0, upper_bound=9999):
+        """Filter by column values in [lower_bound, upper_bound)"""
+        self.df = self.df[(self.df[column] >= lower_bound)
+                          & (self.df[column] < upper_bound)]
+        return self
+
+    def has_tags(self, invert_filter=False):
+        """Select all entries with tags"""
+        mask = self.df["tags"].astype(bool)
+        if invert_filter:
+            mask = ~mask
+        self.df = self.df[mask]
+        return self
+
+    def tags_include(self, tags: Union[List[str], str],
+                     include_any: bool = False):
+        """Select all entries with all/one of given tags
+
+        Arguments:
+        tags: converted to unary list if it's a single string
+        include_any: whether filtering will select entries with just some of
+                     the tags
+        """
+        selector = any if include_any else all
+        def filter_fn(row): return selector(t in row["tags"] for t in tags)
+
+        if type(tags) == str:
+            tags = [tags]
+
+        filter_column = self.df.apply(filter_fn, axis=1)
+        self.df = self.df[filter_column]
+        return self
+
+    def has_comments(self):
+        """Select all entries with comments"""
+        self.df = self.df[self.df["comments"].astype(bool)]
+        return self
+
+    def date(self, lower_bound='1990-01-01', upper_bound='2100-01-01'):
+        """Filter by date in format YYYY-MM-DD"""
+        self.df = self.df[(self.df['date'] >= lower_bound)
+                          & (self.df['date'] < upper_bound)]
+        return self
+
+    def last_x_days(self, x=90):
+        """Select all entries in the last x days"""
+        # TODO pegar data recente
+        now = pd.Timestamp.now()
+        delta = pd.Timedelta(-x, 'd')
+        self.df = self.df[self.df["date"] > now + delta]
+        return self
