--- src/tools_aibg/explorer/__init__.py
@@ -1,1 +1,0 @@
-from .explorer import Explorer
