--- src/tools_aibg/explorer/loaders/__init__.py
@@ -1,1 +1,0 @@
-from .diaguard_import_reader import DiaguardImportReader
