--- src/tools_aibg/reporters/base_reporter/__init__.py
@@ -1,1 +1,0 @@
-from .reporter import Reporter
