--- src/tools_aibg/reporters/base_reporter/analyser/__init__.py
@@ -1,1 +1,0 @@
-from .analyser import Analyser
